Add Tiered Compilation to JitBench
This adds a second set of measurements to the JitBench scenario test with COMPLUS_EXPERIMENTAL_TieredCompilation=1.
I also added a new "--setupjitbench False" argument that helped me test my change more quickly. Its a quick and dirty thing so if anyone doesn't like it we can delete that part, but I don't want spend effort improving it further when it appears all this code will be soon replaced.